move useSearchParamsState to g2f-dashboard
