+++ src/components/pages/dma.tsx
@@ -7,7 +7,7 @@ import { ChartRaceBareDMA } from "../chart_racebar_dma";
 
 import alasql from "alasql";
 import ChartPieTypeTraitement from "../chart_pie_type_traitement";
-import { useSearchParamsState } from "../../utils";
+import { useSearchParamsState } from "../../g2f-dashboard/utils/useSearchParamsState";
 import { DashboardElement } from "../../g2f-dashboard/components/dashboard_element";
