+++ src/components/pages/rep.tsx
@@ -1,6 +1,6 @@
 import { BaseRecord, IResourceComponentsProps, useList } from "@refinedev/core";
 import { Row, Col, Card, List } from "antd";
-import { useSearchParamsState } from "../../utils";
+import { useSearchParamsState } from "../../g2f-dashboard/utils/useSearchParamsState";
 import { ChartPieRepCollecte } from "../chart_pie_rep_collecte";
 import { LoadingComponent } from "../../g2f-dashboard/components/loading_container";
 import { Attribution } from "../../g2f-dashboard/components/attributions";
