+++ src/components/pages/rep_dispmed.tsx
@@ -1,5 +1,5 @@
 import { BaseRecord, IResourceComponentsProps, useList } from "@refinedev/core"
-import { useSearchParamsState } from "../../utils"
+import { useSearchParamsState } from "../../g2f-dashboard/utils/useSearchParamsState"
 import { Row, Col, Card } from "antd"
 import { Attribution } from "../../g2f-dashboard/components/attributions"
 import { ChartPieRepCollecte } from "../chart_pie_rep_collecte"
