+++ src/components/pages/rep_mnu.tsx
@@ -1,5 +1,5 @@
 import { BaseRecord, IResourceComponentsProps, useList } from "@refinedev/core"
-import { useSearchParamsState } from "../../utils"
+import { useSearchParamsState } from "../../g2f-dashboard/utils/useSearchParamsState"
 import { Row, Col, Card } from "antd"
 import { Attribution } from "../../g2f-dashboard/components/attributions"
 import { LoadingComponent } from "../../g2f-dashboard/components/loading_container"
